refactor(frontend): Move request from components to libs
Fixes #5
