+++ frontend/src/stores/mission.ts
@@ -1,4 +1,4 @@
-import { get, post } from "@/components/request/http";
+import { get, post } from "@/libs/request/http";
 import { defineStore } from "pinia";
 import { mainStore } from './main';
 import { FlightMission } from "./models";
